+++ tests/test_stat_result.py
@@ -5,7 +5,7 @@ from statannotations.stats.StatResult import StatResult
 
 
 class TestStatResult(unittest.TestCase):
-    """Test  correction implementation."""
+    """Test `statannotations.stats.StatResult` class"""
 
     def setUp(self) -> None:
         self.benjamini_hochberg = ComparisonsCorrection("Benjamini-Hochberg")