+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## v3
-### v3.0 (In development)
+### v3.0
 (Explicitly Requires Python 3.6)
 
- - Refactoring with implementation of `StatTest` ([#7](https://github.com/trevismd/statannotations/pull/5) in preparation)
+ - Refactoring with implementation of `StatTest`
+   ([#7](https://github.com/trevismd/statannotations/pull/5)), removing the 
+   `stat_func` parameter, and `test_long_name`.
  - Annotations y-positions based on plot coordinates instead of data coordinates 
    ([#5](https://github.com/trevismd/statannotations/pull/5) by [JosephLalli](https://github.com/JosephLalli), fixes https://github.com/webermarcolivier/statannot/issues/21).
  - Add this CHANGELOG